Comentário e exclusão

+++ app/Http/Controllers/UserController.php
@@ -323,23 +323,6 @@ class UserController extends Controller
                                 return view('/login')->withmessage('Erro! Falha na atualização!');
                         }
 
-                        //dd($request->file('teste'));
-
-                        $file = $request->image;
-
-                        // Se informou o arquivo, retorna um boolean
-                        $file = $request->hasFile('image');
-                        
-                        // Se é válido, retorna um boolean
-                        $file = $request->hasFile('image');
-                        
-
-                        $upload = $request->image->storeAs('temp', $request->file('image')->getClientOriginalName());
-
-                        if ( !$upload ){
-                                dd($upload);
-                        }
-
                         $retorno = '/perfil_informacoes';
 
                         $data = $request->only(['name', 'sobrenome', 'email']);
@@ -361,9 +344,11 @@ class UserController extends Controller
 
                         $user->fill($data);
 
-                        Storage::delete("temp/{$user->foto_usuario}");
-
-                        $user->foto_usuario = $request->file('image')->getClientOriginalName();
+                        if ($request->hasFile('image')) {
+                                $upload = $request->image->storeAs('temp', $request->file('image')->getClientOriginalName());
+                                Storage::delete("temp/{$user->foto_usuario}");
+                                $user->foto_usuario = $request->file('image')->getClientOriginalName();
+                        }
 
                         $user->save();
